update: dspy_postprocess_outputs + refactor integration utils
